feature: split workflow instance and task statemachine
