+++ scaleph-workflow/scaleph-workflow-api/pom.xml
@@ -31,14 +31,6 @@
     <dependencies>
         <dependency>
             <groupId>${project.parent.groupId}</groupId>
-            <artifactId>scaleph-system</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>${project.parent.groupId}</groupId>
-            <artifactId>scaleph-queue</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>${project.parent.groupId}</groupId>
             <artifactId>scaleph-dag</artifactId>
         </dependency>
         <dependency>
@@ -61,11 +53,6 @@
         <dependency>
             <groupId>org.apache.pekko</groupId>
             <artifactId>pekko-serialization-jackson_${scala.binary.version}</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>com.alibaba.cola</groupId>
-            <artifactId>cola-component-statemachine</artifactId>
         </dependency>
 
         <dependency>
