+++ scaleph-workflow/scaleph-workflow-simple/pom.xml
@@ -26,18 +26,21 @@
         <version>2.0.4-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
-    <artifactId>scaleph-workflow-quartz</artifactId>
+    <artifactId>scaleph-workflow-simple</artifactId>
 
     <dependencies>
         <dependency>
             <groupId>${project.parent.groupId}</groupId>
-            <artifactId>scaleph-workflow-simple</artifactId>
+            <artifactId>scaleph-workflow-api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>${project.parent.groupId}</groupId>
+            <artifactId>scaleph-queue</artifactId>
         </dependency>
 
         <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-quartz</artifactId>
+            <groupId>com.alibaba.cola</groupId>
+            <artifactId>cola-component-statemachine</artifactId>
         </dependency>
     </dependencies>
-
 </project>
